Synced-From: bazelbuild/buildtools/build Fix end statement tracking (#193)
End statement tracking wasn't accurate and didn't always work well, but on the other hand it was used only for implicit ; insertion (which is not required for Skylark syntax, removing an explicit ; from the end of a line never changes Skylark semantics).

However proper end statement tracking is crucial for bzl files parsing: files are not just sequences of top-level statements, it's important to see where each node ends.

Statements are now considered ended before \n of their last line, and the \n is now passed to the parser explicitly.

GitOrigin-RevId: 15fb4dc16c9c1e6f9e65653bbfea5e056bfe9795