Synced-From: google/starlark-go/syntax compile: optimize a+b statically for literal string/list/tuple (#97)
* compile: optimize a+b statically for literal string/list/tuple

Previously, the parser would fold two string literals "a"+"b" into a
fake string literal, to avoid quadratic runtime cost when a long chain
of strings is added (as often happens due to line length limits).
This made the syntax tree unfaithful to the source,
preventing the frontend from being used in a reformatting tool.

This change does the same optimization in the compiler,
and generalizes it to list and tuple expressions too:

   "a  b" + "c  d" => "a  bc  d"
   [a, b] + [c, d] => [a, b, c, d]
   (a, b) + (c, d) => (a, b, c, d)

It also creates the seam for later addition of the "accumulator"
optimization, in which n-ary non-literal additions such as a+b+c are
optimized to:
   acc = start(a)
   acc.add(b)
   acc.add(c)
   acc.done()
I evaluated this optimization but it doesn't yet pull its weight
on the kinds of builds I can do so far.

GitOrigin-RevId: 60e4b3d65da9be653c231e535f7e1e56d47ca617