Synced-From: bazelbuild/buildtools/build Don't format non-tuple sequences as implicit tuples (#257)
Implicit tuples (e.g. a, b = c) used to be detected based on the fact that the line number of opening brackets for such tuples is 0 which can't be true if a bracket has been parsed from a real file. However, a tuple generated by a refactoring tool always has zeroed positions, meaning that their line numbers for all elements including the brackets are all 0. This commit resolves the ambiguity by using an explicit boolean flag for bracketless tuples.

GitOrigin-RevId: d28382b20795d4cdff1fa280597fea4bd780b758