Synced-From: google/starlark-go/syntax starlark: return EvalErr for errors originating in built-ins (#188)
Previously, only errors originating in Starlark code would be
represented as EvalErr. The backtrace would include frames for
built-in functions that the exception passed through, but not
for the originating frame.

Also, report "<builtin>" not "<builtin>:1" for the position
of such frames. Position.Line==0 no longer implies !IsValid.
Clients cannot rely on valid position strings containing a colon.

Added test of errors originating in a built-in frame.

GitOrigin-RevId: 2494ae9765bca224fde085d23e5d87ed5168fc43