Synced-From: google/starlark-go/syntax starlark: fix bug in int(string, base=int) (#344)
Previously, when int was called with an explicit base,
it would report an error if the digit string starts
with a base prefix for a different base, such as int("0b101", 16).
Now, it uses the base prefix only if it matches the requested
base, so the example above would return 0x0b101, as would
int("0x0b101", 16).

The int(string, int) case has been split out for clarity.

Update doc.

Fixes #337

GitOrigin-RevId: 28488fade247cf7c918bd994f16b1f38b4b7c5bc